Intitial Commit - Pressure Field 3D Task 5
Pressure Field 3D calculates the pressure field due to a point source and returns a 3D matrix, a rank 3 tensor with pressure values in a spatial 2D matrix and the third dimension being time. This is the initial commit and this function is made for Task 5, computing the field of a point source and displaying it as an image.