Initial Commit - Pressure Field 4D Task 6
Pressure Field 4D is a function to calculate the pressure field of a line source (multiple point sources) and it is initially aimed for it to return a 4D matrix or rank 4 tensor. So the pressure intensity being a 2x2 matrix for different x and y values in the observation plane, third dimension being time, and fourth dimension being the original position of the point source in the xs direction of the source plane. Currently, it only returns a 3D matrix (rank 3 tensor), disregarding the position of the source, but needs optimising so that it successfully returns a 4D tensor.